Extract cli args definitions from client (#92)
I did this now because I can see that for #88, I will reuse the same cli-args several times.

In my opinion, this increases readability. This does not include every cli arg yet. I left some that are only used once in the client due to time constraints, and I first want to feel if it is even a good approach. 

After finishing #88, I will formulate a proper client refactoring task, which can be done step-wise.