[client] Add remote ipfs functionality for bazaar (#102)
Uploading files to ipfs remote functionality added,
If flag -l or --ipfs-local is set, the local ipfs is used, else remote.
No more folder upload. 
Icons path as zip file uploaded, each business and offering as single json file added.

ipfs-api-key and the ipfs add to node api url are set as environment varibles. 

When getting file the filename is the hash, is it required to get the original name?